No search results state for reader
